Switch to my puppet-sahara implementation for the time being.
