+++ sahara/.fixtures.yml
@@ -6,9 +6,7 @@ fixtures:
       repo: 'git://github.com/puppetlabs/puppetlabs-mysql.git'
       ref: 'origin/2.2.x'
     'openstacklib': 'git://github.com/stackforge/puppet-openstacklib.git'
-    'postgresql':
-      repo: "git://github.com/puppetlabs/puppet-postgresql.git"
-      ref: '2.5.0'
+    'postgresql': 'git://github.com/puppetlabs/puppetlabs-postgresql.git'
     'qpid': 'git://github.com/dprince/puppet-qpid.git'
     'rabbitmq':
       repo: 'git://github.com/puppetlabs/puppetlabs-rabbitmq'
