+++ sahara/Gemfile
@@ -2,7 +2,10 @@ source 'https://rubygems.org'
 
 group :development, :test do
   gem 'puppetlabs_spec_helper', :require => false
-  gem 'puppet-lint', '~> 0.3.2'
+  gem 'puppet-lint', '~> 1.1'
+  gem 'puppet-lint-param-docs', '1.1.0'
+  gem 'metadata-json-lint'
+  gem 'puppet-syntax'
   gem 'rake', '10.1.1'
   gem 'rspec', '< 2.99'
   gem 'json'
@@ -14,5 +17,3 @@ if puppetversion = ENV['PUPPET_GEM_VERSION']
 else
   gem 'puppet', :require => false
 end
-
-# vim:ft=ruby
