--- sahara/README.rst
@@ -1,1 +1,0 @@
-This repo will contain puppet manifests for Sahara
