+++ sahara/manifests/init.pp
@@ -1,109 +1,172 @@
-# Copyright 2013 Zuercher Hochschule fuer Angewandte Wissenschaften
-# All Rights Reserved.
-#
-#    Licensed under the Apache License, Version 2.0 (the "License"); you may
-#    not use this file except in compliance with the License. You may obtain
-#    a copy of the License at
-#
-#         http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#    License for the specific language governing permissions and limitations
-#    under the License.
-
 # == Class: sahara
 #
-# Installs the sahara backend.
+#  Sahara base package & configuration
 #
 # === Parameters
 #
-# Document parameters here.
+# [*package_ensure*]
+#   (Optional) Ensure state for package
+#   Defaults to 'present'.
 #
-# [*sample_parameter*]
-#   Explanation of what this parameter affects and what it defaults to.
-#   e.g. "Specify one or more upstream ntp servers as an array."
+# [*manage_service*]
+#   (optional) Whether the service should be managed by Puppet.
+#   Defaults to true.
 #
-# === Variables
+# [*enabled*]
+#   (optional) Should the service be enabled.
+#   Defaults to true.
 #
-# Here you should define a list of variables that this module would require.
+# [*verbose*]
+#   (Optional) Should the daemons log verbose messages
+#   Defaults to 'false'.
 #
-# [*sahara_host*]
-# The host on which the sahara process (API) runs. Defaults to '127.0.0.1'
-# [*sahara_port*]
-# The port on which the sahara process (API) runs on. Defaults to 3836
-# [*db_host*]
-# The host where the database is running. Sahara will use this to persist
-# information about clusters. Defaults to '127.0.0.1'
-# [*sahara_db_name*]
-# [*sahara_db_password*]
-# [*keystone_auth_protocol*]
-# Defaults to 'http',
-# [*keystone_auth_host*]
-# Defaults to '127.0.0.1'
-# [*keystone_auth_port*]
-# Defaults to '35357'
-# [*keystone_user*]
-# Defaults to 'sahara'
-# [*keystone_password*]
-# Defaults to 'sahara'
-# [*keystone_tenant*]
-# Defaults to undef
-# [*sahara_verbose*]
-# Defaults to false
-# [*sahara_debug*]
-# Defaults to false
-# === Examples
+# [*debug*]
+#   (Optional) Should the daemons log debug messages
+#   Defaults to 'false'.
 #
-# class{'sahara':
-#   sahara_host              => '127.0.0.1',
-#   db_host                   => '127.0.0.1',
-#   sahara_db_password       => 'sahara',
-#   keystone_auth_host        => '127.0.0.1',
-#   keystone_password         => 'admin',
-#   sahara_verbose           => True,
-#}
+# [*service_host*]
+#   (Optional) Hostname for sahara to listen on
+#   Defaults to '127.0.0.1'.
 #
-# === Authors
+# [*service_port*]
+#   (Optional) Port for sahara to listen on
+#   Defaults to 8386.
 #
-# Andy Edmonds <andy@edmonds.be>
+# [*use_neutron*]
+#   (Optional) Whether to use neutron
+#   Defaults to true.
 #
+# [*database_connection*]
+#   (Optional) Non-sqllite database for sahara
+#   Defaults to 'mysql://sahara:secrete@localhost:3306/sahara'.
 #
-# TODOs
-# - need to install disk builder and create image
-#   or generate and install
-# - use a puppet type for configuration file
-# - clean up documentation
-
-class sahara (
-  $sahara_host           = '127.0.0.1',
-  $sahara_port           = '8386',
-  $sahara_verbose        = false,
-  $sahara_debug          = false,
-  # db
-  $db_host                = '127.0.0.1',
-  $sahara_db_name        = 'sahara',
-  $sahara_db_user        = 'sahara',
-  $sahara_db_password    = 'sahara',
-  # keystone
-  $keystone_auth_protocol = 'http',
-  $keystone_auth_host     = '127.0.0.1',
-  $keystone_auth_port     = '35357',
-  $keystone_user          = 'sahara',
-  $keystone_password      = 'sahara',
-  $keystone_tenant        = undef,) {
+# == keystone authentication options
+#
+# [*os_username*]
+#   (Optional) Username for sahara credentials
+#   Defaults to 'admin'.
+#
+# [*os_password*]
+#   (Optional) Password for sahara credentials
+#   Defaults to 'secrete'.
+#
+# [*os_tenant_name*]
+#   (Optional) Tenant for os_username
+#   Defaults to 'admin'.
+#
+# [*os_auth_url*]
+#   (Optional) Public identity endpoint
+#   Defaults to 'http://127.0.0.1:5000/v2.0/'.
+#
+# [*identity_url*]
+#   (Optional) Admin identity endpoint
+#   Defaults to 'http://127.0.0.1:35357/'.
+#
+class sahara(
+  $manage_service      = true,
+  $enabled             = true,
+  $package_ensure      = 'present',
+  $verbose             = false,
+  $debug               = false,
+  $service_host        = '127.0.0.1',
+  $service_port        = 8386,
+  $use_neutron         = true,
+  $database_connection = 'mysql://sahara:secrete@localhost:3306/sahara',
+  $os_username         = 'admin',
+  $os_password         = 'secrete',
+  $os_tenant_name      = 'admin',
+  $os_auth_url         = 'http://127.0.0.1:5000/v2.0/',
+  $identity_url        = 'http://127.0.0.1:35357/',
+) {
   include sahara::params
 
-  # move keystone and db classes here?
-
-  if !$keystone_tenant {
-    $int_keystone_tenant = $keystone_user
-  } else {
-    $int_keystone_tenant = $keystone_tenant
+  file { '/etc/sahara/':
+    ensure  => directory,
+    owner   => 'root',
+    group   => 'sahara',
+    mode    => '0750',
+    require => Package['sahara'],
   }
 
-  class { '::sahara::install':
-  } ->
-  class { '::sahara::service': }
+  file { '/etc/sahara/sahara.conf':
+    owner   => 'root',
+    group   => 'sahara',
+    mode    => '0640',
+    require => File['/etc/sahara'],
+  }
+
+  package { 'sahara':
+    ensure => $package_ensure,
+    name   => $::sahara::params::package_name,
+  }
+
+  Package['sahara'] -> Sahara_config<||>
+  Package['sahara'] ~> Service['sahara']
+
+  validate_re($database_connection, '(sqlite|mysql|postgresql):\/\/(\S+:\S+@\S+\/\S+)?')
+
+  case $database_connection {
+    /^mysql:\/\//: {
+      require mysql::bindings
+      require mysql::bindings::python
+    }
+    /^postgresql:\/\//: {
+      require postgresql::lib::python
+    }
+    /^sqlite:\/\//: {
+      fail('Sahara does not support sqlite!')
+    }
+    default: {
+      fail('Unsupported db backend configured')
+    }
+  }
+
+  sahara_config {
+    'DEFAULT/use_neutron': value => $use_neutron;
+    'DEFAULT/host': value => $service_host;
+    'DEFAULT/port': value => $service_port;
+    'DEFAULT/debug': value => $debug;
+    'DEFAULT/verbose': value => $verbose;
+
+    'database/connection':
+      value => $database_connection,
+      secret => true;
+
+    'keystone_authtoken/auth_uri': value => $os_auth_url;
+    'keystone_authtoken/identity_uri': value => $identity_url;
+    'keystone_authtoken/admin_user': value => $os_username;
+    'keystone_authtoken/admin_tenant_name': value => $os_tenant_name;
+    'keystone_authtoken/admin_password':
+      value => $os_password,
+      secret => true;
+  }
+
+  if $manage_service {
+    if $enabled {
+      $service_ensure = 'running'
+    } else {
+      $service_ensure = 'stopped'
+    }
+  }
+
+  Package['sahara'] -> Service['sahara']
+  service { 'sahara':
+    ensure     => $service_ensure,
+    name       => $::sahara::params::service_name,
+    hasstatus  => true,
+    enable     => $enabled,
+    hasrestart => true,
+    subscribe  => Exec['sahara-dbmanage'],
+  }
+
+  Sahara_config<||> ~> Exec['sahara-dbmanage']
+
+  exec { 'sahara-dbmanage':
+    command     => $::sahara::params::dbmanage_command,
+    path        => '/usr/bin',
+    user        => 'root',
+    refreshonly => true,
+    logoutput   => on_failure,
+  }
+
 }
