+++ sahara/manifests/keystone/auth.pp
@@ -1,73 +1,125 @@
-# Copyright 2013 Zuercher Hochschule fuer Angewandte Wissenschaften
-# All Rights Reserved.
+# == Class: sahara::keystone::auth
 #
-#    Licensed under the Apache License, Version 2.0 (the "License"); you may
-#    not use this file except in compliance with the License. You may obtain
-#    a copy of the License at
+# Configures sahara service and endpoint in Keystone.
 #
-#         http://www.apache.org/licenses/LICENSE-2.0
+# === Parameters
 #
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#    License for the specific language governing permissions and limitations
-#    under the License.
+# [*password*]
+#   (required) Password for Sahara user.
+#
+# [*service_name*]
+#   (Optional) Name of the service.
+#   Defaults to the value of auth_name.
+#
+# [*auth_name*]
+#   (Optional) Username for sahara service.
+#   Defaults to 'sahara'.
+#
+# [*email*]
+#   (Optional) Email for Sahara user.
+#   Defaults to 'sahara@localhost'.
+#
+# [*tenant*]
+#   (Optional) Tenant for Sahara user.
+#   Defaults to 'services'.
+#
+# [*configure_endpoint*]
+#   (Optional) Should Sahara endpoint be configured?
+#   Defaults to 'true'.
+#
+# [*service_type*]
+#   (Optional) Type of service.
+#   Defaults to 'data_processing'.
+#
+# [*service_description*]
+#   (Optional) Description of service.
+#   Defaults to 'Sahara Data Processing'.
+#
+# [*region*]
+#   (Optional) Region for endpoint.
+#   Defaults to 'RegionOne'.
+#
+# [*public_protocol*]
+#   (Optional) Protocol for public endpoint.
+#   Defaults to 'http'.
+#
+# [*admin_protocol*]
+#   (Optional) Protocol for admin endpoint.
+#   Defaults to 'http'.
+#
+# [*internal_protocol*]
+#   (Optional) Protocol for internal endpoint.
+#   Defaults to 'http'.
+#
+# [*public_address*]
+#   (Optional) Public address for endpoint.
+#   Defaults to '127.0.0.1'.
+#
+# [*admin_address*]
+#   (Optional) Admin address for endpoint.
+#   Defaults to '127.0.0.1'.
+#
+# [*internal_address*]
+#   (Optional) Internal address for endpoint.
+#   Defaults to '127.0.0.1'.
+#
+# [*port*]
+#   (Optional) Port for endpoint.
+#   Defaults to '8386'.
+#
+# [*public_port*]
+#   Port for public endpoint. Defaults to $port.
+#
+# [*version*]
+#   (Optional) Version of API to use.
+#   Defaults to 'v1.1'.
+#
+class sahara::keystone::auth(
+  $password,
+  $service_name        = undef,
+  $auth_name           = 'sahara',
+  $email               = 'sahara@localhost',
+  $tenant              = 'services',
+  $service_type        = 'data_processing',
+  $service_description = 'Sahara Data Processing',
+  $configure_endpoint  = true,
+  $region              = 'RegionOne',
+  $public_protocol     = 'http',
+  $admin_protocol      = 'http',
+  $internal_protocol   = 'http',
+  $public_address      = '127.0.0.1',
+  $admin_address       = '127.0.0.1',
+  $internal_address    = '127.0.0.1',
+  $port                = '8386',
+  $public_port         = undef,
+  $version             = 'v1.1',
+) {
 
-#
-# Used to setup the sahara keystone user
-#
+  if $service_name == undef {
+    $real_service_name = $auth_name
+  } else {
+    $real_service_name = $service_name
+  }
 
-class sahara::keystone::auth (
-  $password = 'sahara',
-  $auth_name = 'sahara',
-  $email = 'sahara@localhost',
-  $tenant = 'services',
-  $configure_endpoint = true,
-  $service_type = 'data_processing',
-  $public_address = '127.0.0.1',
-  $admin_address = '127.0.0.1',
-  $internal_address = '127.0.0.1',
-  $port = '8386',
-  $public_port = undef,
-  $region = 'RegionOne',
-  $public_protocol = 'http',
-  $internal_protocol = 'http',
-  ) {
-
-  Keystone_user_role["${auth_name}@${tenant}"] ~>
-    Service <| name == 'sahara-api' |>
-
-  if !$public_port {
+  if ! $public_port {
     $real_public_port = $port
   } else {
     $real_public_port = $public_port
   }
 
-  keystone_user { $auth_name:
-    ensure   => present,
-    password => $password,
-    email    => $email,
-    tenant   => $tenant,
-  }
 
-  keystone_user_role { "${auth_name}@${tenant}":
-    ensure => present,
-    roles  => 'admin',
-  }
-
-  keystone_service { $auth_name:
-    ensure      => present,
-    type        => $service_type,
-    description => 'Openstack Data Processing',
-  }
-
-  if $configure_endpoint {
-    keystone_endpoint { "${region}/${auth_name}":
-      ensure       => present,
-      public_url   =>
-        "${public_protocol}://${public_address}:${real_public_port}/",
-      internal_url => "${internal_protocol}://${internal_address}:${port}/",
-      admin_url    => "${internal_protocol}://${admin_address}:${port}/",
-    }
+  keystone::resource::service_identity { $real_service_name:
+    configure_user      => true,
+    configure_user_role => true,
+    configure_endpoint  => $configure_endpoint,
+    service_type        => $service_type,
+    service_description => $service_description,
+    region              => $region,
+    password            => $password,
+    email               => $email,
+    tenant              => $tenant,
+    public_url          => "${public_protocol}://${public_address}:${real_public_port}/${version}/%(tenant_id)s",
+    admin_url           => "${admin_protocol}://${admin_address}:${port}/${version}/%(tenant_id)s",
+    internal_url        => "${internal_protocol}://${internal_address}:${port}/${version}/%(tenant_id)s",
   }
 }
