+++ sahara/metadata.json
@@ -1,33 +1,37 @@
 {
   "name": "stackforge-sahara",
-  "version": "4.0.0",
-  "author": "StackForge Contributors",
+  "version": "5.0.0",
+  "author": "Red Hat and StackForge Contributors",
   "summary": "Puppet module for OpenStack Sahara",
   "license": "Apache License 2.0",
-  "source": "git://github.com/stackforge/puppet-sahara.git",
-  "project_page": "https://launchpad.net/puppet-sahara",
-  "issues_url": "https://bugs.launchpad.net/puppet-sahara",
+  "source": "git://github.com/frozencemetery/puppet-sahara.git",
+  "project_page": "https://github.com/frozencemetery/puppet-sahara",
+  "issues_url": "https://github.com/frozencemetery/puppet-sahara/issues",
   "requirements": [
-    { "name": "pe","version_requirement": "3.2.x" },
+    { "name": "pe","version_requirement": "3.x" },
     { "name": "puppet","version_requirement": "3.x" }
   ],
   "operatingsystem_support": [
     {
-      "operatingsystem": "Debian",
-      "operatingsystemrelease": ["7"]
+      "operatingsystem": "Fedora",
+      "operatingsystemrelease": ["20"]
     },
     {
       "operatingsystem": "RedHat",
-      "operatingsystemrelease": ["6","7"]
+      "operatingsystemrelease": ["7"]
     },
     {
-      "operatingsystem": "Ubuntu",
-      "operatingsystemrelease": ["14.04"]
+      "operatingsystem": "Debian",
+      "operatingsystemrelease": ["8"]
     }
   ],
-  "description": "Installs and configures OpenStack Sahara (Data Processing).",
+  "description": "Installs and configures OpenStack Sahara (Hadoop).",
   "dependencies": [
+    { "name": "duritong/sysctl", "version_requirement": ">=0.0.1 <1.0.0" },
+    { "name": "stackforge/keystone", "version_requirement": ">=5.0.0 <6.0.0" },
+    { "name": "puppetlabs/inifile", "version_requirement": ">=5.0.0 <6.0.0" },
     { "name": "puppetlabs/stdlib", "version_requirement": ">=4.0.0 <5.0.0" },
+    { "name": "puppetlabs/postgresql", "version_requirement": ">=3.0.0" },
     { "name": "stackforge/openstacklib", "version_requirement": ">=5.0.0" }
   ]
 }
