--- sahara/templates/sahara-api-redhat.erb
@@ -1,102 +1,0 @@
-#!/bin/sh
-#
-# sahara-api OpenStack Hadoop Cluster API
-#
-# chkconfig:   - 99 02
-# description: OpenStack Hadoop Cluster Management API service
-
-### BEGIN INIT INFO
-# Provides:
-# Required-Start: $remote_fs $network $syslog
-# Required-Stop: $remote_fs $syslog
-# Default-Stop: 0 1 6
-# Short-Description: Sahara API server
-# Description: OpenStack Hadoop Cluster Management API service
-### END INIT INFO
-
-. /etc/rc.d/init.d/functions
-
-suffix=api
-prog=sahara-$suffix
-config="/etc/sahara/sahara.conf"
-exec="/usr/bin/sahara-$suffix"
-pidfile="/var/run/sahara/sahara-$suffix.pid"
-user=sahara
-
-[ -e /etc/sysconfig/$prog ] && . /etc/sysconfig/$prog
-
-lockfile=/var/lock/subsys/$prog
-
-start() {
-    [ -x $exec ] || exit 5
-    [ -f $config ] || exit 6
-    echo -n $"Starting $prog: "
-    daemon --user $user --pidfile $pidfile "$exec -d &>/dev/null & echo \$! > $pidfile"
-    retval=$?
-    echo
-    [ $retval -eq 0 ] && touch $lockfile
-    return $retval
-}
-
-stop() {
-    echo -n $"Stopping $prog: "
-    killproc -p $pidfile $prog
-    retval=$?
-    echo
-    [ $retval -eq 0 ] && rm -f $lockfile
-    return $retval
-}
-
-restart() {
-    stop
-    start
-}
-
-reload() {
-    restart
-}
-
-force_reload() {
-    restart
-}
-
-rh_status() {
-    status $prog
-}
-
-rh_status_q() {
-    rh_status >/dev/null 2>&1
-}
-
-
-case "$1" in
-    start)
-        rh_status_q && exit 0
-        $1
-        ;;
-    stop)
-        rh_status_q || exit 0
-        $1
-        ;;
-    restart)
-        $1
-        ;;
-    reload)
-        rh_status_q || exit 7
-        $1
-        ;;
-    force-reload)
-        force_reload
-        ;;
-    status)
-        rh_status
-        ;;
-    condrestart|try-restart)
-        rh_status_q || exit 0
-        restart
-        ;;
-    *)
-        echo $"Usage: $0 {start|stop|status|restart|condrestart|try-restart|reload|force-reload}"
-        exit 2
-esac
-exit $?
