--- sahara/templates/sahara-api.conf.erb
@@ -1,18 +1,0 @@
-description "Sahara API Server upstart job"
-author "Andy Edmonds <andy@edmonds.be>"
-
-start on runlevel [2345]
-stop on runlevel [!2345]
-
-
-chdir <%= scope.lookupvar("sahara::params::sys_rundir") %>
-
-pre-start script
-    mkdir -p <%= scope.lookupvar("sahara::params::sahara_rundir") %>
-    chown sahara:root <%= scope.lookupvar("sahara::params::sahara_rundir") %>
-
-    mkdir -p <%= scope.lookupvar("sahara::params::sahara_lockdir") %>
-    chown sahara:root <%= scope.lookupvar("sahara::params::sahara_lockdir") %>
-end script
-
-exec start-stop-daemon --start --chuid sahara --exec /usr/local/bin/<%= scope.lookupvar("sahara::params::sahara_service") %> -- --config-file=<%= scope.lookupvar("sahara::params::sahara_conf_file") %>
